requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
SQLAlchemy==2.0.28
asyncpg==0.29.0
APScheduler==3.10.4 